ddtracer/tracer: move AdaptLogger and related code into tracer package
